+++ crates/core/tests/java/align_multiline_text_blocks/block_align.out.java
@@ -1,0 +1,8 @@
+class A {
+    void m() {
+        String s = """
+                hello
+                  world
+                """;
+    }
+}