+++ crates/core/tests/java/strip_whitespace_from_blank_lines_in_text_blocks/block.java
@@ -1,0 +1,11 @@
+class A {
+    void m() {
+        String s = """
+        alpha
+         
+        beta
+        	
+        gamma
+        """;
+    }
+}